+++ packages/terra-form-input/CHANGELOG.md
@@ -4,6 +4,18 @@ ChangeLog
 Unreleased
 ----------
 
+2.21.0 - (July 30, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+* Updated errorIcon default prop from `Field.defaultProps.errorIcon` with `<IconError>` component.
+* updated package.json test scripts
+
+2.20.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.19.0 - (July 17, 2019)
 ------------------
 ## Changed